requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
requests
insightface
onnxruntime
opencv-python-headless
numpy
pillow
tqdm
multipart